LPS-53578 rename package
