+++ modules/apps/site-navigation/site-navigation-site-map-web/src/com/liferay/site/navigation/site/map/web/configuration/SiteMapWebConfigurationUtil.java
@@ -12,16 +12,22 @@
  * details.
  */
 
-package com.liferay.site.map.web.configuration;
+package com.liferay.site.navigation.site.map.web.configuration;
 
-import com.liferay.portal.kernel.util.GetterUtil;
+import com.liferay.portal.kernel.configuration.Configuration;
+import com.liferay.portal.kernel.configuration.ConfigurationFactoryUtil;
 
 /**
  * @author Eudaldo Alonso
  */
-public class SiteMapWebConfigurationValues {
+public class SiteMapWebConfigurationUtil {
 
-	public static final String DISPLAY_TEMPLATES_CONFIG = GetterUtil.getString(
-		SiteMapWebConfigurationUtil.get("display.templates.config"));
+	public static String get(String key) {
+		return _configuration.get(key);
+	}
+
+	private static final Configuration _configuration =
+		ConfigurationFactoryUtil.getConfiguration(
+			SiteMapWebConfigurationUtil.class.getClassLoader(), "portlet");
 
 }
